requirements.txt
langchain>=0.0.205
openai>=0.27.8 
tqdm>=4.65.0
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0
pytest>=7.3.2
streamlit_chat
pgvector
psycopg2-binary
tiktoken